# 08-rest/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Library API
  description: This is a simple API for managing books in a library.
  version: 1.0.0
servers:
  - url: 'http://api.yourlibrary.com/v1'
paths:
  /books:
    get:
      summary: List all books
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      summary: Add a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Book created

  /books/{id}:
    get:
      summary: Get a book by its ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    put:
      summary: Update a book by its ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book updated
    delete:
      summary: Delete a book by its ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Book deleted

components:
  schemas:
    Book:
      type: object
      required:
        - title
        - author
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        description:
          type: string
          nullable: true