# 08-rest/src/main/java/rest/com/Book.java
package rest.com;

public class Book {
    private int id;
    private String title;
    private String author;
    private String description;
    
    public Book(String title, String author, String description) {
        this.title = title;
        this.author = author;
        this.description = description;
    }
	public String getTitle() {
		return title;
	}
	public void setTitle(String title) {
		this.title = title;
	}
	public int getId() {
		return id;
	}
	public void setId(int id) {
		this.id = id;
	}
	public String getDescription() {
		return description;
	}
	public void setDescription(String description) {
		this.description = description;
	}
	public String getAuthor() {
		return author;
	}
	public void setAuthor(String author) {
		this.author = author;
	}

}